added AbstractTransformTableProcessor
