+++ data-processing-lib/src/data_processing/runtime/__init__.py
@@ -1,2 +1,3 @@
 from data_processing.runtime.execution_configuration import TransformExecutionConfiguration
 from data_processing.runtime.transform_launcher import AbstractTransformLauncher
+from data_processing.runtime.transform_table_processor import AbstractTransformTableProcessor
